add a wrapper for concept type and return THING if concept type is not set
This safe guard is to cover the case where a user upload a document or video and the GPW didn't get a chance to set a concept type before a user tries to publish.

Currently our backend will error out with a NPE. Adding a wrapper like how the front end does it where if you ask us for a concept type and the property doesn't exist we will return THING instead of null.